+++ CHANGELOG.rst
@@ -6,6 +6,26 @@ All notable changes to this project will be documented in this file.
 .. contents:: Table of Contents
    :depth: 2
 
+v2.1.0 - 2024-08-18
+===================
+
+Added
+-----
+
+* Support for Django 4.2
+* Allow forms to be overridden from settings
+
+Deprecated
+----------
+
+* Make the crypt module optional and deprecate it's usage.
+  The python stdlid crypt module is deprecated since version 3.11 and
+  will be removed in version 3.13. Check for the availability of the
+  crypt module.
+  All password checks using the crypt module will stop to work on
+  python 3.13.
+
+
 v2.0.0 - 2022-10-17
 ===================